+++ src/Components/SingUp/SingUp.css
@@ -1,0 +1,52 @@
+.form {
+    display: flex;
+    padding: 1% 6%;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.letter {
+    font-weight: bold;
+}
+
+.card {
+    width: 50%;
+    border: 2px solid gray;
+    border-radius: 15px;
+    background-color: lightgray;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    overflow: hidden;
+    transition: all 300ms;
+    cursor: pointer;
+    box-shadow: 0 10px 50px rgba(0, 0, 0, 0.3);
+    height: 500px;
+}
+
+.card:hover {
+    transform: scale(1.02) translateY(-15px);
+}
+
+ul {
+    list-style: none;
+}
+
+li span {
+    text-decoration: none;
+    text-transform: uppercase;
+    color: blue;
+    margin-right: 5px;
+}
+
+.card-header {
+    height: 300px;
+    width: 100%;
+    overflow: hidden;
+}
+
+.card-header img {
+    width: 100%;
+    height: 100%;
+}
